No setup
Program and tape are now local variables.
There is no more setup because of that, input and output are accessible
attributes while check_program is up to the user to call when required.
Ruby mode had WRITE bug fixed and can use evaluation mode to access
external input/output, which allows better testing.